Split ipc-handlers.ts into domain-specific modules
Refactor the monolithic ipc-handlers.ts (490+ lines, 29 handlers) into
6 focused modules under src/main/ipc/:
- pipeline-ipc.ts: pipeline start/stop, engine switching
- settings-ipc.ts: settings load/save, subtitle settings, glossary
- session-ipc.ts: session management, export, summary generation
- display-ipc.ts: display enumeration, subtitle window
- audio-ipc.ts: WebSocket audio server
- model-ipc.ts: model status queries, plugins, GPU detection

The main ipc-handlers.ts becomes a thin coordinator that delegates
to each domain module's register function.
